Move salesforcedx-vscode-apex-debugger tests to vscode-integration dir
